Move signals out of models #50
