+++ mailqueue/apps.py
@@ -5,3 +5,6 @@ class MailQueueConfig(AppConfig):
     name = 'mailqueue'
     label = 'mailqueue'
     verbose_name = "Mail Queue"
+
+    def ready(self):
+        from mailqueue.receivers import *
