+++ mailqueue/models.py
@@ -8,8 +8,6 @@ from django.utils.translation import ugettext_lazy as _
 from django.utils import timezone
 from django.db import models
 from django.core.mail import EmailMultiAlternatives
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from django.conf import settings
 
 from . import defaults
@@ -132,15 +130,3 @@ class Attachment(models.Model):
 
     def __str__(self):
         return self.file_attachment.name
-
-
-@receiver(post_save, sender=MailerMessage)
-def send_post_save(sender, instance, signal, *args, **kwargs):
-    if getattr(instance, "do_not_send", False):
-        instance.do_not_send = False
-        return
-
-    if not getattr(settings, 'MAILQUEUE_QUEUE_UP', defaults.MAILQUEUE_QUEUE_UP):
-        # If mail queue up is set, wait for the cron or management command
-        # to send any email.
-        instance.send_mail()
