+++ pytest.ini
@@ -9,5 +9,6 @@ flakes-ignore =
     docs/*.py ALL
     mailqueue/tests/test_smoke.py ALL
     mailqueue/south_migrations/*.py ALL
+    mailqueue/apps.py ImportStarUsed
 
 pep8maxlinelength = 100
